refactor: diagram canvas
