+++ package.json
@@ -36,7 +36,7 @@
     "@types/react-router-dom": "^5.1.0",
     "@types/react-test-renderer": "^16.9.2",
     "@types/recharts": "^1.1.21",
-    "@types/styled-components": "^4.4.0",
+    "@types/styled-components": "^5.1.9",
     "@typescript-eslint/eslint-plugin": "^2.7.0",
     "@typescript-eslint/parser": "^2.7.0",
     "@webpack-cli/serve": "^1.3.1",
@@ -68,7 +68,7 @@
     "react-test-renderer": "^17.0.1",
     "recharts": "^1.7.1",
     "style-loader": "^1.1.3",
-    "styled-components": "^5.0.0",
+    "styled-components": "^5.2.1",
     "ts-jest": "^26.5.5",
     "ts-loader": "^6.2.2",
     "typescript": "^4.2.4",
